Revert "Webserv encode"
