+++ cliapp-cust/go.mod
@@ -1,0 +1,3 @@
+module myapp
+
+go 1.18
