+++ cliapp-cust/myapp.log
@@ -1,0 +1,6 @@
+Jun  1 22:20:05 secserv kernel: Kernel log (proc) stopped.
+Jun  1 22:20:05 secserv kernel: Kernel ERROR daemon terminating.
+Jun  1 22:20:06 secserv exiting on signal 15
+Nov 27 08:05:57 galileo kernel: Kernel log (proc) stopped.
+Nov 27 08:05:57 galileo kernel: Kernel ERROR daemon terminating.
+Nov 27 08:05:57 galileo exiting on signal 15
