--- webservices-encoder/go.mod
@@ -1,3 +1,0 @@
-module webservices
-
-go 1.18
