+++ webservices-urlquery/go.mod
@@ -1,0 +1,3 @@
+module webservices
+
+go 1.18
